***** fdutils version 5.5 diff of 2008/10/27 *****
The last version that was delivered with Slackware
from 14.1 and up to 14.2.

Changes since 5.5:

v20081027
  * Add "first-sector-number" flag.

v20060227
  * Fixed "zero-based" flag

Originated-by: fdutils-5.5.tar.gz
Originated-md5: 2b8b65f52378158a4a21e455566b456d
Originated-URL: http://www.fdutils.linux.lu/fdutils-5.5.tar.gz

Originated-by: fdutils-5.5-20081027.diff.gz
Originated-md5: 5babc93625d13a860d1b4a93bf24020b
Originated-URL: http://www.fdutils.linux.lu/fdutils-5.5-20081027.diff.gz
Originated-URL: https://mirrors.slackware.com/slackware/slackware-14.2/source/a/floppy/fdutils-5.5-20081027.diff.gz

+++ src/floppycontrol.c
@@ -379,7 +379,7 @@ int main( int argc, char **argv)
 	}
 
 	if (mask & SET_RESET)
-		eioctl(fd, FDRESET, (void *)reset_now, "reset");
+		eioctl(fd, FDRESET, (void *)(long)reset_now, "reset");
 
 	if (mask & (PRINTSTATE | POLLSTATE) ){
 		if ( mask & POLLSTATE )
